changed the balance from BigDecimal to int due to the project requirements

+++ src/main/java/com/example/domain/Account.java
@@ -1,10 +1,8 @@
 package com.example.domain;
 
-import java.math.BigDecimal;
-
 public interface Account {
 
     Card getCard();
 
-    BigDecimal getBalance();
+    int getBalance();
 }
